fix: memory safety, resource leaks, save-data loss and parity gaps
Code review of ~1M lines. Verified against a clean-tree baseline: 47 of 558
targeted tests fail without these changes, 43 with them, zero regressions.

PROVEN (test fails without the fix, passes with it):
* GAME_LOAD read only the first element of each fixed globals section --
  4 of 200 bytes -- leaving the SUPPRESS stream 1568 bits short, so heroes,
  savegames1, timers and the record-link stream decoded from the wrong bit
  offset. sksvgame.cpp:1512-1514 reads (1 x 8), (1 x 0x40), (2 x 0x40) and
  the reader indexes mask[i] for i < count, so each section must be read per
  element. No existing test caught it: GAME_LOAD's own reader.position check
  compares against an offset derived from the same read sequence. Adds the
  missing writer/reader cross-check.
* Parry was never subtracted from incoming creature melee (PROJEXPL.C:1392
  subtracts parry << 1). DM1_ChampionCombat had no skillParry field at all.
* handle_mask was 32-bit while item_class spans 0..0x3d, so the shift was
  undefined and bit 9 silently answered for the moneybox class 41.

MEMORY SAFETY (each traced to a concrete triggering input): int overflow
removing the bound on a 64 KB stack buffer in the GRAPHICS.DAT reader;
strcpy of a 1024-byte line into 512 bytes in the shipped asset scanner (x2);
unbounded IMG9-LZW chain walk plus a dictionary memset covering 256 of 4096
entries; one-past-the-end write into decode_stack; OOB read and directory
cycle in the FM Towns disc parser; unbounded ISO directory recursion
reachable from any user disc image; FAT12 index past its buffer; strcpy
starting past the terminator; MVE size_t underflow; uint32 overflow.

RESOURCE LEAKS: fclose() inside short-circuiting || chains at 15 sites. The
worst leaked a FILE* and its fd, then unlinked the still-open file.

SAVE DATA LOSS: csb_v1_save_backup silently dropped the ".backup" suffix for
long paths -- truncating the live save it was meant to protect -- and
ignored every write error while reporting success, after which the caller
overwrote the original. Plus a double fclose on a failed deferred flush.

NO FABRICATED CONTENT: generated sine/noise effects replaced missing source
SND3; V2.2 drew every creature as a demon and every wall and door as one
carved stone; DM1 painted an invented dialog plaque where only CSB failed
closed. All now leave source pixels intact or stay silent.

PARITY: push/pull never turned the party; dropped items lost their tile
quadrant; mouse turning was dead (missing idx -= 79); the rune panel had no
mana half; moneybox flag; bar remainder segment.

TESTS: dungeon_loader_fmtowns asserted cwTextData (28) as the map count --
nMaps is 44, confirmed three ways (DME.h:93 offset 4, 0x2c in both PC and FM
Towns headers, descriptors 0..43 well-formed with the column index at
44 + 44*16 = 748). creature_combat_probe segfaulted on strcmp(NULL, ...) and
demanded 30 AI names; ai_name and sound_name deliberately return NULL
because those labels have no source text owner.

KNOWN GAPS documented in code rather than left silently wrong:
DM2_ChampionRecord offsets do not match the on-disk record; the orchestrator
does not own the possession-link array; player_testing_wall is a rewrite not
a port; the two dungeon-loader variants disagree on header offsets (the
pc_g1 layout is required by its fixtures, so it is documented not changed).

+++ include/dm2_v1_runtime_narrow_pc34_compat.h
@@ -688,7 +688,11 @@ int32_t dm2_v1_creature_attacks_creature(
  * (flags & 0x80, i.e. a forced pickpocket/steal), otherwise the regular
  * per-type handle_mask bit for the item's class decides. */
 typedef struct {
-    uint32_t handle_mask;   /* bit N set => creature type can handle class N */
+    /* bit N set => creature type can handle class N.
+     * 64-bit because item_class is (item & 0x3f), i.e. 0..0x3d: classes 32
+     * and above -- including the moneybox class 0x29 -- are not
+     * representable in a 32-bit mask, and shifting by them is undefined. */
+    uint64_t handle_mask;
 } DM2_V1_CreatureHandleCaps;
 
 typedef struct {
